refactor: 카드 등록 페이지 컴포넌트 폴더 구조 생성
