+++ src/components/CardRegistrationPage/FormContents/ExpirationDateInput.tsx
@@ -1,8 +1,8 @@
-import Input from "../common/Input";
-import InputBox from "../common/InputBox";
-import InputGroup from "../common/InputGroup";
-import InputSeparator from "../common/InputSeparator";
-import { useCardItemAction, useCardItemValue, useErrorMessageValue } from "../provider/CardItemProvider";
+import Input from "../../common/Input";
+import InputBox from "../../common/InputBox";
+import InputGroup from "../../common/InputGroup";
+import InputSeparator from "../../common/InputSeparator";
+import { useCardItemAction, useCardItemValue, useErrorMessageValue } from "../../provider/CardItemProvider";
 
 const ExpirationDateInput = () => {
   const { expirationDate } = useCardItemValue();
