+++ src/components/CardRegistrationPage/FormContents/PasswordInput.tsx
@@ -1,9 +1,9 @@
 import styled from "styled-components";
-import Input from "../common/Input";
-import InputBox from "../common/InputBox";
-import InputGroup from "../common/InputGroup";
-import { DotIcon } from "../../assets/icons";
-import { useCardItemAction, useCardItemValue, useErrorMessageValue } from "../provider/CardItemProvider";
+import Input from "../../common/Input";
+import InputBox from "../../common/InputBox";
+import InputGroup from "../../common/InputGroup";
+import { DotIcon } from "../../../assets/icons";
+import { useCardItemAction, useCardItemValue, useErrorMessageValue } from "../../provider/CardItemProvider";
 
 const PasswordInput = () => {
   const { password } = useCardItemValue();
