+++ src/components/CardRegistrationPage/FormContents/SecurityCodeInput.tsx
@@ -1,7 +1,7 @@
-import Input from "../common/Input";
-import InputBox from "../common/InputBox";
-import InputGroup from "../common/InputGroup";
-import { useCardItemAction, useCardItemValue, useErrorMessageValue } from "../provider/CardItemProvider";
+import Input from "../../common/Input";
+import InputBox from "../../common/InputBox";
+import InputGroup from "../../common/InputGroup";
+import { useCardItemAction, useCardItemValue, useErrorMessageValue } from "../../provider/CardItemProvider";
 
 const SecurityCodeInput = () => {
   const { securityCode } = useCardItemValue();
